refactor: extract AI trace store from ai_trace handlers (NEU-539)
Move every VictoriaLogs interaction — LogsQL query construction, the
NDJSON decode shapes, and value quoting — into a new traceStore type
(trace_store.go). Handlers now deal in whole AITrace values and HTTP
concerns only; all store queries grow from baseQuery(), giving the
upcoming chunked-body support a single place to hook record filtering
and body reassembly.

Pure refactor: query strings and wire behavior are unchanged.
