+++ internal/routes/logs/ai_trace.go
@@ -1,11 +1,8 @@
 package logs
 
 import (
-	"bufio"
-	"encoding/json"
 	"fmt"
 	"net/http"
-	"net/url"
 	"strconv"
 	"strings"
 	"time"
@@ -45,23 +42,9 @@ type AITrace struct {
 	ResponseBody     string `json:"response_body,omitempty"`
 }
 
-// listProjection is the LogsQL `fields` projection for the list query: every
-// metadata column the list view renders, deliberately excluding the large
-// request_body / response_body fields so list responses stay small.
-const listProjection = "_time, request_id, workspace, endpoint_type, " +
-	"endpoint_name, api_key_id, request_uri, request_model, response_model, " +
-	"response_status, prompt_tokens, completion_tokens, total_tokens, " +
-	"finish_reason, stream, user_agent, duration_ms"
-
 // stringTrue is the literal a boolean-valued string (a query flag, or a
 // stringified VictoriaLogs field) must equal to be considered on.
 const stringTrue = "true"
-
-// fullProjection extends listProjection with the large request/response body
-// columns. The list endpoint uses it only when the caller passes
-// ?include_body=true — chiefly the CLI export, which fetches bodies inline to
-// avoid an N+1 per-record detail lookup.
-const fullProjection = listProjection + ", request_body, response_body"
 
 // AITraceListResponse is the wire format for GET /api/v1/ai-traces/:workspace.
 //
@@ -83,6 +66,25 @@ type AITraceDayCount struct {
 // the activity bar chart at the top of the SPA's trace list.
 type AITraceStatsResponse struct {
 	Days []AITraceDayCount `json:"days"`
+}
+
+// AITraceKeyStat is one API key's aggregated traffic over the requested window.
+// Powers the API-key list ranking overview and the detail "request performance"
+// card. SuccessRate is derived client-side as Success/Requests.
+type AITraceKeyStat struct {
+	APIKeyID      string  `json:"api_key_id"`
+	Requests      int64   `json:"requests"`
+	Tokens        int64   `json:"tokens"`
+	Success       int64   `json:"success"`
+	AvgDurationMs float64 `json:"avg_duration_ms"`
+}
+
+// AITraceKeyStatsResponse is the wire format for
+// GET /api/v1/ai-traces/:workspace/key-stats — per-API-key request count, token
+// total, success count and average latency over a trailing window (default 24h).
+type AITraceKeyStatsResponse struct {
+	WindowHours int              `json:"window_hours"`
+	Keys        []AITraceKeyStat `json:"keys"`
 }
 
 // RegisterAITraceRoutes mounts the AI inference trace endpoints.
@@ -342,13 +344,23 @@ func traceEndpointTypeFilter(c *gin.Context) string {
 	return fmt.Sprintf("endpoint_type:=%s", logsQLQuoteValue(et))
 }
 
+// requireTraceStore resolves the trace store, or answers 503 and returns nil
+// when no --ai-trace-store-url is configured.
+func requireTraceStore(c *gin.Context, deps *Dependencies) *traceStore {
+	store := newTraceStore(deps)
+	if store == nil {
+		c.JSON(http.StatusServiceUnavailable, gin.H{
+			"error": "ai trace store is not configured",
+		})
+	}
+
+	return store
+}
+
 func handleListAITraces(deps *Dependencies) gin.HandlerFunc {
 	return func(c *gin.Context) {
-		if deps.AITraceStoreURL == "" {
-			c.JSON(http.StatusServiceUnavailable, gin.H{
-				"error": "ai trace store is not configured",
-			})
-
+		store := requireTraceStore(c, deps)
+		if store == nil {
 			return
 		}
 
@@ -360,65 +372,31 @@ func handleListAITraces(deps *Dependencies) gin.HandlerFunc {
 			}
 		}
 
-		// Leading clause scopes the query to the workspace(s) the caller may read
-		// (a concrete workspace, or the cross-workspace "_all_" aggregate).
-		queryParts := []string{traceScopeClause(c)}
-
-		if endpoint := strings.TrimSpace(c.Query("endpoint_name")); endpoint != "" {
-			queryParts = append(queryParts, fmt.Sprintf("endpoint_name:=%s", logsQLQuoteValue(endpoint)))
+		filters := traceFilters{
+			EndpointName: strings.TrimSpace(c.Query("endpoint_name")),
+			EndpointType: strings.TrimSpace(c.Query("endpoint_type")),
+			Status:       strings.TrimSpace(c.Query("status")),
+			APIKeyID:     strings.TrimSpace(c.Query("api_key_id")),
+			FinishReason: strings.TrimSpace(c.Query("finish_reason")),
+			Model:        strings.TrimSpace(c.Query("model")),
 		}
 
-		if endpointType := strings.TrimSpace(c.Query("endpoint_type")); endpointType != "" {
-			queryParts = append(queryParts, fmt.Sprintf("endpoint_type:=%s", logsQLQuoteValue(endpointType)))
-		}
+		// The list view never renders request/response bodies, so they are
+		// omitted by default. Callers exporting the full record opt in via
+		// ?include_body=true so bodies come back inline, avoiding a per-record
+		// detail lookup.
+		includeBody := c.Query("include_body") == stringTrue
 
-		if status := strings.TrimSpace(c.Query("status")); status != "" {
-			queryParts = append(queryParts, fmt.Sprintf("response_status:=%s", logsQLQuoteValue(status)))
-		}
-
-		if apiKeyID := strings.TrimSpace(c.Query("api_key_id")); apiKeyID != "" {
-			queryParts = append(queryParts, fmt.Sprintf("api_key_id:=%s", logsQLQuoteValue(apiKeyID)))
-		}
-
-		if fr := strings.TrimSpace(c.Query("finish_reason")); fr != "" {
-			queryParts = append(queryParts, fmt.Sprintf("finish_reason:=%s", logsQLQuoteValue(fr)))
-		}
-
-		if model := strings.TrimSpace(c.Query("model")); model != "" {
-			// model can match either request or response model
-			queryParts = append(queryParts, fmt.Sprintf(
-				"(request_model:=%s OR response_model:=%s)",
-				logsQLQuoteValue(model), logsQLQuoteValue(model),
-			))
-		}
-
-		// The list view never renders request/response bodies — project them
-		// out so VictoriaLogs returns only the small metadata columns. Callers
-		// exporting the full record (?include_body=true) opt into the larger
-		// projection so bodies come back inline, avoiding a per-record lookup.
-		projection := listProjection
-		if c.Query("include_body") == stringTrue {
-			projection = fullProjection
-		}
-
-		query := strings.Join(queryParts, " ") +
-			" | sort by (_time) desc | limit " + strconv.Itoa(limit) +
-			" | fields " + projection
-
-		params := url.Values{}
-
-		if start := strings.TrimSpace(c.Query("start")); start != "" {
-			params.Set("start", start)
-		}
+		window := timeWindow{Start: strings.TrimSpace(c.Query("start"))}
 		// `before` is the inclusive upper bound from the previous page's last
 		// record's timestamp. To page strictly older we pass it as end.
 		if before := strings.TrimSpace(c.Query("before")); before != "" {
-			params.Set("end", before)
-		} else if end := strings.TrimSpace(c.Query("end")); end != "" {
-			params.Set("end", end)
+			window.End = before
+		} else {
+			window.End = strings.TrimSpace(c.Query("end"))
 		}
 
-		items, err := queryAITraces(deps, query, params)
+		items, err := store.List(traceScopeClause(c), filters, limit, includeBody, window)
 		if err != nil {
 			klog.Errorf("ai-trace: list: %v", err)
 			c.JSON(http.StatusBadGateway, gin.H{
@@ -445,11 +423,8 @@ func handleListAITraces(deps *Dependencies) gin.HandlerFunc {
 // so the SPA's detail drawer fetches them lazily through this route.
 func handleGetAITrace(deps *Dependencies) gin.HandlerFunc {
 	return func(c *gin.Context) {
-		if deps.AITraceStoreURL == "" {
-			c.JSON(http.StatusServiceUnavailable, gin.H{
-				"error": "ai trace store is not configured",
-			})
-
+		store := requireTraceStore(c, deps)
+		if store == nil {
 			return
 		}
 
@@ -465,12 +440,7 @@ func handleGetAITrace(deps *Dependencies) gin.HandlerFunc {
 		// The scope clause confines the lookup to workspaces/endpoint types the
 		// caller may read, so an "_all_" detail fetch cannot read a trace the
 		// caller has no permission for.
-		query := fmt.Sprintf(
-			"%s request_id:=%s | sort by (_time) desc | limit 1",
-			traceScopeClause(c), logsQLQuoteValue(requestID),
-		)
-
-		items, err := queryAITraces(deps, query, url.Values{})
+		item, err := store.Get(traceScopeClause(c), requestID)
 		if err != nil {
 			klog.Errorf("ai-trace: detail: %v", err)
 			c.JSON(http.StatusBadGateway, gin.H{
@@ -480,7 +450,7 @@ func handleGetAITrace(deps *Dependencies) gin.HandlerFunc {
 			return
 		}
 
-		if len(items) == 0 {
+		if item == nil {
 			c.JSON(http.StatusNotFound, gin.H{
 				"error": "trace not found",
 			})
@@ -488,7 +458,7 @@ func handleGetAITrace(deps *Dependencies) gin.HandlerFunc {
 			return
 		}
 
-		c.JSON(http.StatusOK, items[0])
+		c.JSON(http.StatusOK, item)
 	}
 }
 
@@ -500,11 +470,8 @@ const maxStatsDays = 90
 // powering the activity bar chart at the top of the SPA's trace list.
 func handleAITraceStats(deps *Dependencies) gin.HandlerFunc {
 	return func(c *gin.Context) {
-		if deps.AITraceStoreURL == "" {
-			c.JSON(http.StatusServiceUnavailable, gin.H{
-				"error": "ai trace store is not configured",
-			})
-
+		store := requireTraceStore(c, deps)
+		if store == nil {
 			return
 		}
 
@@ -517,19 +484,10 @@ func handleAITraceStats(deps *Dependencies) gin.HandlerFunc {
 			return
 		}
 
-		query := fmt.Sprintf(
-			"%s | stats by (_time:1d) count() total",
-			traceScopeClause(c),
-		)
-
-		params := url.Values{}
-		params.Set("start", startDay.Format(time.RFC3339))
 		// `end` is pushed to the start of the day after endDay so the final
 		// (possibly partial) day is always fully covered, regardless of whether
 		// VictoriaLogs treats the bound as inclusive or exclusive.
-		params.Set("end", endDay.AddDate(0, 0, 1).Format(time.RFC3339))
-
-		counts, err := queryAITraceDayCounts(deps, query, params)
+		counts, err := store.DayCounts(traceScopeClause(c), startDay, endDay.AddDate(0, 0, 1))
 		if err != nil {
 			klog.Errorf("ai-trace: stats: %v", err)
 			c.JSON(http.StatusBadGateway, gin.H{
@@ -619,77 +577,6 @@ func statsWindow(c *gin.Context) (startDay, endDay time.Time, ok bool) {
 	return startDay, endDay, true
 }
 
-// queryAITraceDayCounts runs a `stats by (_time:1d)` LogsQL query and returns
-// a map of UTC date (YYYY-MM-DD) to request count.
-func queryAITraceDayCounts(deps *Dependencies, query string, params url.Values) (map[string]int, error) {
-	params.Set("query", query)
-	reqURL := strings.TrimRight(deps.AITraceStoreURL, "/") + "/select/logsql/query?" + params.Encode()
-
-	resp, err := deps.HTTPClient.Get(reqURL)
-	if err != nil {
-		return nil, fmt.Errorf("query victorialogs: %w", err)
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != http.StatusOK {
-		return nil, fmt.Errorf("victorialogs returned status %d", resp.StatusCode)
-	}
-
-	counts := make(map[string]int)
-	scanner := bufio.NewScanner(resp.Body)
-
-	for scanner.Scan() {
-		line := scanner.Bytes()
-		if len(line) == 0 {
-			continue
-		}
-
-		var r struct {
-			Time  string `json:"_time"`
-			Total string `json:"total"`
-		}
-
-		if err := json.Unmarshal(line, &r); err != nil {
-			continue
-		}
-
-		// `_time` is an RFC3339 day bucket; key by the date component.
-		date := r.Time
-		if len(date) >= 10 {
-			date = date[:10]
-		}
-
-		if n, err := strconv.Atoi(r.Total); err == nil {
-			counts[date] = n
-		}
-	}
-
-	if err := scanner.Err(); err != nil {
-		return nil, fmt.Errorf("scan victorialogs response: %w", err)
-	}
-
-	return counts, nil
-}
-
-// AITraceKeyStat is one API key's aggregated traffic over the requested window.
-// Powers the API-key list ranking overview and the detail "request performance"
-// card. SuccessRate is derived client-side as Success/Requests.
-type AITraceKeyStat struct {
-	APIKeyID      string  `json:"api_key_id"`
-	Requests      int64   `json:"requests"`
-	Tokens        int64   `json:"tokens"`
-	Success       int64   `json:"success"`
-	AvgDurationMs float64 `json:"avg_duration_ms"`
-}
-
-// AITraceKeyStatsResponse is the wire format for
-// GET /api/v1/ai-traces/:workspace/key-stats — per-API-key request count, token
-// total, success count and average latency over a trailing window (default 24h).
-type AITraceKeyStatsResponse struct {
-	WindowHours int              `json:"window_hours"`
-	Keys        []AITraceKeyStat `json:"keys"`
-}
-
 // maxKeyStatsWindowHours bounds the trailing window the key-stats endpoint will
 // aggregate over (30 days), keeping a single LogsQL scan bounded.
 const maxKeyStatsWindowHours = 720
@@ -701,11 +588,8 @@ const maxKeyStatsWindowHours = 720
 // one call (the detail view picks its key out of the result).
 func handleAITraceKeyStats(deps *Dependencies) gin.HandlerFunc {
 	return func(c *gin.Context) {
-		if deps.AITraceStoreURL == "" {
-			c.JSON(http.StatusServiceUnavailable, gin.H{
-				"error": "ai trace store is not configured",
-			})
-
+		store := requireTraceStore(c, deps)
+		if store == nil {
 			return
 		}
 
@@ -717,21 +601,9 @@ func handleAITraceKeyStats(deps *Dependencies) gin.HandlerFunc {
 			}
 		}
 
-		// Success = 2xx/3xx response_status (regex match, robust to the field being
-		// stored as a string); tokens sums total_tokens (missing => 0); avg latency
-		// over duration_ms. Empty api_key_id rows (untagged traffic) are dropped by
-		// the parser.
-		query := fmt.Sprintf(
-			"%s | stats by (api_key_id) count() requests, "+
-				"sum(total_tokens) tokens, avg(duration_ms) avg_duration_ms, "+
-				"count() if (response_status:~\"^[23]\") success",
-			traceScopeClause(c),
-		)
+		since := time.Now().UTC().Add(-time.Duration(windowHours) * time.Hour)
 
-		params := url.Values{}
-		params.Set("start", time.Now().UTC().Add(-time.Duration(windowHours)*time.Hour).Format(time.RFC3339))
-
-		keys, err := queryAITraceKeyStats(deps, query, params)
+		keys, err := store.KeyStats(traceScopeClause(c), since)
 		if err != nil {
 			klog.Errorf("ai-trace: key-stats: %v", err)
 			c.JSON(http.StatusBadGateway, gin.H{
@@ -746,245 +618,4 @@ func handleAITraceKeyStats(deps *Dependencies) gin.HandlerFunc {
 			Keys:        keys,
 		})
 	}
-}
-
-// queryAITraceKeyStats runs the per-key `stats by (api_key_id)` aggregation and
-// decodes the NDJSON result rows (all values arrive as strings from VL).
-func queryAITraceKeyStats(deps *Dependencies, query string, params url.Values) ([]AITraceKeyStat, error) {
-	params.Set("query", query)
-	reqURL := strings.TrimRight(deps.AITraceStoreURL, "/") + "/select/logsql/query?" + params.Encode()
-
-	resp, err := deps.HTTPClient.Get(reqURL)
-	if err != nil {
-		return nil, fmt.Errorf("query victorialogs: %w", err)
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != http.StatusOK {
-		return nil, fmt.Errorf("victorialogs returned status %d", resp.StatusCode)
-	}
-
-	out := make([]AITraceKeyStat, 0, 16)
-	scanner := bufio.NewScanner(resp.Body)
-
-	for scanner.Scan() {
-		line := scanner.Bytes()
-		if len(line) == 0 {
-			continue
-		}
-
-		var r struct {
-			APIKeyID      string `json:"api_key_id"`
-			Requests      string `json:"requests"`
-			Tokens        string `json:"tokens"`
-			AvgDurationMs string `json:"avg_duration_ms"`
-			Success       string `json:"success"`
-		}
-
-		if err := json.Unmarshal(line, &r); err != nil {
-			continue
-		}
-
-		// Drop the untagged bucket (requests with no api_key_id): it cannot be
-		// attributed to any key in the ranking.
-		if strings.TrimSpace(r.APIKeyID) == "" {
-			continue
-		}
-
-		// VL may format count()/sum() results as plain ints or floats
-		// ("1240000" or "1.24e6"); parse as float and truncate for robustness.
-		stat := AITraceKeyStat{
-			APIKeyID:      r.APIKeyID,
-			Requests:      parseIntLoose(r.Requests),
-			Tokens:        parseIntLoose(r.Tokens),
-			Success:       parseIntLoose(r.Success),
-			AvgDurationMs: parseFloatLoose(r.AvgDurationMs),
-		}
-
-		out = append(out, stat)
-	}
-
-	if err := scanner.Err(); err != nil {
-		return nil, fmt.Errorf("scan victorialogs response: %w", err)
-	}
-
-	return out, nil
-}
-
-// parseIntLoose parses a VL numeric result (which may be int- or float-
-// formatted) into an int64, truncating any fractional part. Returns 0 on error.
-// Integer-formatted values parse as base-10 int64 first so large counts
-// (>= 2^53) keep full precision; only float/scientific forms fall back to float.
-func parseIntLoose(s string) int64 {
-	s = strings.TrimSpace(s)
-	if n, err := strconv.ParseInt(s, 10, 64); err == nil {
-		return n
-	}
-
-	if f, err := strconv.ParseFloat(s, 64); err == nil {
-		return int64(f)
-	}
-
-	return 0
-}
-
-// parseFloatLoose parses a VL numeric result into a float64, returning 0 on error.
-func parseFloatLoose(s string) float64 {
-	if f, err := strconv.ParseFloat(strings.TrimSpace(s), 64); err == nil {
-		return f
-	}
-
-	return 0
-}
-
-// queryAITraces runs a LogsQL query against VictoriaLogs and decodes the
-// NDJSON response into AITrace records.
-func queryAITraces(deps *Dependencies, query string, params url.Values) ([]AITrace, error) {
-	params.Set("query", query)
-	reqURL := strings.TrimRight(deps.AITraceStoreURL, "/") + "/select/logsql/query?" + params.Encode()
-
-	resp, err := deps.HTTPClient.Get(reqURL)
-	if err != nil {
-		return nil, fmt.Errorf("query victorialogs: %w", err)
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != http.StatusOK {
-		return nil, fmt.Errorf("victorialogs returned status %d", resp.StatusCode)
-	}
-
-	items := make([]AITrace, 0, 64)
-	scanner := bufio.NewScanner(resp.Body)
-	// A detail record can be large because it embeds the full request and
-	// response bodies.
-	scanner.Buffer(make([]byte, 64*1024), 16*1024*1024)
-
-	for scanner.Scan() {
-		line := scanner.Bytes()
-		if len(line) == 0 {
-			continue
-		}
-
-		item, ok := decodeVLRecord(line)
-		if !ok {
-			continue
-		}
-
-		items = append(items, item)
-	}
-
-	if err := scanner.Err(); err != nil {
-		return nil, fmt.Errorf("scan victorialogs response: %w", err)
-	}
-
-	return items, nil
-}
-
-// vlRecord matches the shape Vector writes to VictoriaLogs.
-// All values come back as strings; we coerce types we care about.
-type vlRecord struct {
-	Time             string `json:"_time"`
-	Stream           string `json:"_stream,omitempty"`
-	RequestID        string `json:"request_id"`
-	Workspace        string `json:"workspace"`
-	EndpointType     string `json:"endpoint_type"`
-	EndpointName     string `json:"endpoint_name"`
-	APIKeyID         string `json:"api_key_id"`
-	RequestURI       string `json:"request_uri"`
-	RequestModel     string `json:"request_model"`
-	ResponseModel    string `json:"response_model"`
-	ResponseStatus   string `json:"response_status"`
-	PromptTokens     string `json:"prompt_tokens"`
-	CompletionTokens string `json:"completion_tokens"`
-	TotalTokens      string `json:"total_tokens"`
-	FinishReason     string `json:"finish_reason"`
-	IsStream         string `json:"stream"`
-	UserAgent        string `json:"user_agent"`
-	DurationMs       string `json:"duration_ms"`
-	RequestBody      string `json:"request_body"`
-	ResponseBody     string `json:"response_body"`
-}
-
-func decodeVLRecord(line []byte) (AITrace, bool) {
-	var r vlRecord
-	if err := json.Unmarshal(line, &r); err != nil {
-		return AITrace{}, false
-	}
-
-	t := AITrace{
-		RequestID:     r.RequestID,
-		Time:          r.Time,
-		Workspace:     r.Workspace,
-		EndpointType:  r.EndpointType,
-		EndpointName:  r.EndpointName,
-		APIKeyID:      r.APIKeyID,
-		RequestURI:    r.RequestURI,
-		RequestModel:  r.RequestModel,
-		ResponseModel: r.ResponseModel,
-		FinishReason:  r.FinishReason,
-		Stream:        r.IsStream == stringTrue,
-		UserAgent:     r.UserAgent,
-		RequestBody:   r.RequestBody,
-		ResponseBody:  r.ResponseBody,
-	}
-	if r.Time == "" {
-		t.Time = time.Now().UTC().Format(time.RFC3339Nano)
-	}
-
-	if n, err := strconv.Atoi(r.ResponseStatus); err == nil {
-		t.ResponseStatus = n
-	}
-
-	if r.PromptTokens != "" {
-		if n, err := strconv.Atoi(r.PromptTokens); err == nil {
-			t.PromptTokens = &n
-		}
-	}
-
-	if r.CompletionTokens != "" {
-		if n, err := strconv.Atoi(r.CompletionTokens); err == nil {
-			t.CompletionTokens = &n
-		}
-	}
-
-	if r.TotalTokens != "" {
-		if n, err := strconv.Atoi(r.TotalTokens); err == nil {
-			t.TotalTokens = &n
-		}
-	}
-
-	if r.DurationMs != "" {
-		// `latencies.request` arrives as a float-formatted string from Vector.
-		if f, err := strconv.ParseFloat(r.DurationMs, 64); err == nil {
-			n := int(f)
-			t.DurationMs = &n
-		}
-	}
-
-	return t, true
-}
-
-// logsQLQuoteValue wraps a string literal for an exact-match LogsQL filter
-// (`field:=value`). VL accepts double-quoted strings; inner quotes/backslashes
-// are escaped. Empty input is permitted by the caller and produces `""`.
-func logsQLQuoteValue(v string) string {
-	var b strings.Builder
-	b.Grow(len(v) + 2)
-	b.WriteByte('"')
-
-	for _, r := range v {
-		switch r {
-		case '\\', '"':
-			b.WriteByte('\\')
-			b.WriteRune(r)
-		case '\n':
-			b.WriteString("\\n")
-		default:
-			b.WriteRune(r)
-		}
-	}
-
-	b.WriteByte('"')
-
-	return b.String()
 }
